Use fixed-size types and fix warnings
-Replace most uses of "long" with uint32_t/int32_t

-Replace some uses of "short" with uint16_t/int16_t

-Replace most uses of "char" with uint8_t

This should improve compatibility when compiling
on systems where "long" is a 64-bit size.

+++ n3dout.c
@@ -173,7 +173,7 @@ writeanims(FILE *f, Object *obj)
 {
 	Matrix *M;
 	int i;
-	long x, y, z;
+	int32_t x, y, z;
 
 	fprintf(f, ".%s_anim:\n", name2label(obj->name));
 	fprintf(f, "\t.dc.w\t1, 0\t; frame animation\n");
